CMake: Update lenses handling in Augeas
The same library in two different layers may be structured in two
different ways.

It is then important to keep an interface (in this case the property
named AUGEAS_LENSES_FOLDER_PATH) that is well defined.

The INTERFACE_LIBRARY property used before was assuming that all the
layers contained the lenses in the same relative path, but this is for
example not the case inside the upstream source repository and broke the
build when switching layer.