Add a simple system to patch submodules
This commit adds a system that uses "git apply"
to patch submodules source code.

Patches should be put under libraries/cmake/source/<library name>/patches/<submodule>,
where <submodule> is often one and is "src", but in other cases, like AWS,
there are multiple with a more specific name.

If for whatever reason the submodule cloning or the patching fails,
the submodule has to be unregistered and its folder should be cleared.
This should be achievable with "git submodule deinit -f <submodule path>".

After fixing the patch or the general issue, configuring again should
just work.