Fix googletest cmake project leaking private compiler flags
The minimum CMake version that the googletest project requires is too old
and uses a sometimes incompatible linking system which
with static libraries do not respect the CMake PRIVATE visibility/inheritance.

Save for implementing a custom CMakeLists, the solution for now
is to patch the submodule and increase the minimum required version.