Remove glibc provided symbols and duplicated ones
berkeley-db was implementing a insertElement function which is a copy of
what sqlite already have. Removed that implementation from
the compilation.

libmagic was implementing several standard library functions which are
already present in glibc. Removed those implementations from the
compilation.

lldpd was implementing a setproctitle function which is already provided
by util-linux, while also not using it. Removed that implementation from
the compilation.

util-linux was implementing a crc32 function which is already provided
by zlib. Though since the implementation differs from zlib, to avoid
collisions the symbol and usages have been renamed to utillinux_crc32.
The header where it's declared it's not public anymore.